requirements.txt
geopy
pyswisseph
pytz
